fix(ceda): Test corrections
