+++ src/nwp_consumer/internal/repositories/raw_repositories/__init__.py
@@ -3,14 +3,14 @@
 TODO: Add description
 """
 
-from .ceda_ftp import CEDAFTPRawRepository
+from .ceda import CEDARawRepository
 from .ecmwf_realtime import ECMWFRealTimeS3RawRepository
 from .ecmwf_mars import ECMWFMARSRawRepository
 from .noaa_s3 import NOAAS3RawRepository
 from .mo_datahub import MetOfficeDatahubRawRepository
 
 __all__ = [
-    "CEDAFTPRawRepository",
+    "CEDARawRepository",
     "ECMWFRealTimeS3RawRepository",
     "NOAAS3RawRepository",
     "MetOfficeDatahubRawRepository",
